Add keyboard input

+++ Comp. complexity theory/Sorting/Bubble.h
@@ -7,7 +7,7 @@
 namespace Sorting {
 	class Bubble : public Sort {
 	public:
-		Bubble(const std::vector<int>& numbers) : Sort(numbers) {}
+		Bubble(const std::vector<double>& numbers) : Sort(numbers) {}
 
 		void Sorting() override;
 		void Output() const override;
