+++ Comp. complexity theory/Sorting/Merge.h
@@ -7,7 +7,7 @@
 namespace Sorting {
 	class Merge : public Sort {
 	public:
-		Merge(const std::vector<int>& numbers) : Sort(numbers) {}
+		Merge(const std::vector<double>& numbers) : Sort(numbers) {}
 
 		void Sorting() override;
 		void Output() const override;
