+++ Comp. complexity theory/Sorting/Select.h
@@ -7,7 +7,7 @@
 namespace Sorting {
 	class Select : public Sort {
 	public:
-		Select(const std::vector<int>& numbers) : Sort(numbers) {}
+		Select(const std::vector<double>& numbers) : Sort(numbers) {}
 
 		void Sorting() override;
 		void Output() const override;
